Renamed toxicity_lookup -> toxicreactiontotreatment_lookup to match class name.
